requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
cssselect==1.3.0
cssutils==2.11.1
dill==0.3.9
easydict==1.13
fasttext==0.9.3
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
Jinja2==3.1.6
joblib==1.4.2
klepto==0.2.6
lightning-utilities==0.14.2
lxml==5.3.2
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
networkx==3.2.1
Pillow==9.4.0
pox==0.3.5
premailer==3.10.0
propcache==0.3.0
protobuf==6.30.1
pybind11==2.13.6
pycairo==1.24.0
scikit-learn==1.6.1
scipy==1.13.1
sympy==1.13.1
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
torch-scatter==2.1.2
torchaudio==0.12.1
torchmetrics==1.7.0
torchvision==0.13.1
